+++ src/entities/weekly.entity.ts
@@ -3,9 +3,13 @@ import { WeeklyParticipant } from './weekly-participant.entity';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
+// TODO: Write bot invite guid and configure help command to get the link
 export class Weekly {
     @PrimaryGeneratedColumn()
     public id!: number;
+
+    @Column()
+    public guildId!: string;
 
     @Column({ type: 'timestamp' })
     public date!: Date;